+++ src/utils/sharedFunctions.test.ts
@@ -1,7 +1,7 @@
 import { initGame, Item, Room } from "../gonorth";
 import { goToRoom } from "./lifecycle";
 import { selectPlayer } from "./selectors";
-import { inSameRoomAs } from "./sharedFunctions";
+import { inRoom, inSameRoomAs } from "./sharedFunctions";
 
 jest.mock("./consoleIO");
 const consoleIO = require("./consoleIO");
@@ -29,4 +29,8 @@ describe("inSameRoomAs function", () => {
   it("gives true when player is in room with item", () => expect(inSameRoomAs(ball)).toBe(true));
   it("gives false when player is not in room with item", () => expect(inSameRoomAs(spoon)).toBe(false));
   it("gives true when item is in player's inventory", () => expect(inSameRoomAs(watch)).toBe(true));
+  it("returns true when the player is in the named room", () => expect(inRoom("parlour")).toBe(true));
+  it("returns false when the player is not in the named room", () => expect(inRoom("scullery")).toBe(false));
+  it("returns true when the player is in the named room case notwithstanding", () =>
+    expect(inRoom("PARLOUR")).toBe(true));
 });